Added a lot of features.

+++ src/main/java/com/jeffreymanzione/jef/parsing/IntDefinition.java
@@ -5,7 +5,7 @@ public class IntDefinition extends SingletonDefintion {
 	public static volatile IntDefinition instance;
 
 	private IntDefinition() {
-		super("LONG");
+		super("Int");
 	}
 
 	public static Definition instance() {
